+++ src/Makefile.am
@@ -1,5 +1,5 @@
 lib_LTLIBRARIES = libusbg.la
-libusbg_la_SOURCES = usbg.c
+libusbg_la_SOURCES = usbg.c usbg_schemes_libconfig.c
 libusbg_la_LDFLAGS = $(LIBCONFIG_LIBS)
 libusbg_la_LDFLAGS += -version-info 0:1:0
 libusbg_la_CFLAGS = $(LIBCONFIG_CFLAGS)
